Some more minor fixes for warnings (and errors introduced with the last fix)
* src/folder_browser.cpp
  (TruncatePath): changed type of variable "pos" from "int" to
    "size_t"

* src/diff_dlg.cpp
  (DiffDlg::Data::GetData): added "break" to "default:" statement.

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7045 65390229-12b7-0310-b90b-f21a5aa7ec8e

+++ src/folder_browser.cpp
@@ -94,7 +94,7 @@ TruncatePath (const wxString & path_)
   if (path.length() <= MAXLENGTH_BOOKMARK)
     return path;
 
-  int pos = path.Find (":");
+  size_t pos = path.Find (":");
   wxString newPath;
 
   pos++;
